# TangWebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TangWebApi",
    "description": "基于.NET 8的Web API项目，集成SqlSugar、Swagger和Knife4j",
    "contact": {
      "name": "TangWebApi",
      "email": "admin@tangwebapi.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Cache/set": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "设置缓存",
        "operationId": "SetCache (Cache)",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "缓存键",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "缓存值",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireMinutes",
            "in": "query",
            "description": "过期时间（分钟）",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Cache/get": {
      "get": {
        "tags": [
          "Cache"
        ],
        "summary": "获取缓存",
        "operationId": "GetCache (Cache)",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "缓存键",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Cache/remove": {
      "delete": {
        "tags": [
          "Cache"
        ],
        "summary": "删除缓存",
        "operationId": "RemoveCache (Cache)",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "缓存键",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Cache/exists": {
      "get": {
        "tags": [
          "Cache"
        ],
        "summary": "检查缓存是否存在",
        "operationId": "ExistsCache (Cache)",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "缓存键",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Cache/clear": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "清空所有缓存",
        "operationId": "ClearCache (Cache)",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/Cache/type": {
      "get": {
        "tags": [
          "Cache"
        ],
        "summary": "获取缓存服务类型",
        "operationId": "GetCacheType (Cache)",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "description": "JWT授权token前面需要加上字段Bearer与一个空格,如Bearer token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}